Add unlist built-in to turn the result of head into a regular value
Head returns a Q-Expression but you cannot compare a Q-Expression to a regular number, so unlist takes that number out of the Q-Expression.

An alternative solution would be for head to not return a Q-Expression. In that case you'd need to use list to turn that back into a Q-Expression if so desired.

Also added a number of basic functions written in LISP.